+++ packages/irmin-watcher.0.5.0/src/backend.inotify.mli
@@ -1,35 +1,20 @@
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire. All rights reserved.
    Distributed under the ISC license, see terms at the end of the file.
-   irmin-watcher 0.4.1
+   irmin-watcher 0.5.0
   ---------------------------------------------------------------------------*)
 
-(** Irmin watchers.
+(** Inotify backend for Irmin watchers.
 
-    {e 0.4.1 — {{:https://github.com/mirage/irmin-watcher }homepage}} *)
+    {e 0.5.0 — {{:https://github.com/mirage/irmin-watcher} homepage}} *)
 
-val v: Core.t
-(** [v id p f] is the listen hook calling [f] everytime a sub-path of
-    [p] is modified. Return a function to call to remove the
-    hook. Default to polling if no better solution is
-    available. FSevents and Inotify backends are available. *)
+val v : Core.t Lazy.t
+(** [v id p f] is the hook calling [f] everytime a sub-path of [p] is modified.
+    Return a function to call to remove the hook. Use inofity to be notified on
+    filesystem changes. *)
 
-val mode: [ `FSEvents | `Inotify | `Polling ]
-
-type stats = {
-  watchdogs : int;
-  dispatches: int;
-}
-
-val hook: Core.hook
-(** [hook t] is an {!Irmin.Watcher} compatible representation of {!v}. *)
-
-val stats: unit -> stats
-(** [stats ()] is a snapshot of [v]'s stats. *)
-
-val set_polling_time: float -> unit
-(** [set_polling_time f] set the polling interval to [f]. Only
-    valid when [mode = `Polling]. *)
+val mode : [ `Inotify | `Polling ]
+(** [mode] is [Inotify] on Linux and [`Polling] on Darwin. *)
 
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire