+++ packages/irmin-watcher.0.5.0/src/core.ml
@@ -1,28 +1,29 @@
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire. All rights reserved.
    Distributed under the ISC license, see terms at the end of the file.
-   irmin-watcher 0.4.1
+   irmin-watcher 0.5.0
   ---------------------------------------------------------------------------*)
 
 open Astring
 open Lwt.Infix
 
 let src = Logs.Src.create "irmin-watcher" ~doc:"Irmin watcher logging"
+
 module Log = (val Logs.src_log src : Logs.LOG)
 
 (* run [t] and returns an handler to stop the task. *)
 let stoppable t =
   let s, u = Lwt.task () in
-  Lwt.async (fun () -> Lwt.pick ([s; t ()]));
-  function () -> Lwt.wakeup u (); Lwt.return_unit
+  Lwt.async (fun () -> Lwt.pick [ s; t () ]);
+  function
+  | () ->
+      Lwt.wakeup u ();
+      Lwt.return_unit
 
 external unix_realpath : string -> string = "irmin_watcher_unix_realpath"
 
 let realpath dir =
-  let (/) x y = match y with
-  | None   -> x
-  | Some y -> Filename.concat x y
-  in
+  let ( / ) x y = match y with None -> x | Some y -> Filename.concat x y in
   let rec aux dir file =
     try unix_realpath dir / file
     with Unix.Unix_error (Unix.ENOENT, _, _) ->
@@ -32,35 +33,43 @@ let realpath dir =
   aux dir None
 
 module Digests = struct
-  include Set.Make(struct
-      type t = string * Digest.t
-      let compare = compare
-    end)
+  include Set.Make (struct
+    type t = string * Digest.t
+
+    let compare = compare
+  end)
+
   let of_list l = List.fold_left (fun set elt -> add elt set) empty l
+
   let sdiff x y = union (diff x y) (diff y x)
+
   let digest_pp ppf d = Fmt.string ppf @@ Digest.to_hex d
+
   let pp_elt = Fmt.(Dump.pair string digest_pp)
+
   let pp ppf t = Fmt.(Dump.list pp_elt) ppf @@ elements t
+
   let files t =
     elements t |> List.map fst |> String.Set.of_list |> String.Set.elements
 end
 
 module Dispatch = struct
-
   type t = (string, (int * (string -> unit Lwt.t)) list) Hashtbl.t
 
-  let empty (): t = Hashtbl.create 10
+  let empty () : t = Hashtbl.create 10
+
   let clear t = Hashtbl.clear t
 
-  let stats t ~dir =
-    try List.length (Hashtbl.find t dir) with Not_found -> 0
+  let stats t ~dir = try List.length (Hashtbl.find t dir) with Not_found -> 0
 
   (* call all the callbacks on the file *)
   let apply t ~dir ~file =
     let fns = try Hashtbl.find t dir with Not_found -> [] in
-    Lwt_list.iter_p (fun (id, f) ->
-        Log.debug (fun f -> f "callback %d" id); f file
-      ) fns
+    Lwt_list.iter_p
+      (fun (id, f) ->
+        Log.debug (fun f -> f "callback %d" id);
+        f file)
+      fns
 
   let add t ~id ~dir fn =
     let fns = try Hashtbl.find t dir with Not_found -> [] in
@@ -69,30 +78,22 @@ module Dispatch = struct
 
   let remove t ~id ~dir =
     let fns = try Hashtbl.find t dir with Not_found -> [] in
-    let fns = List.filter (fun (x,_) -> x <> id) fns in
-    if fns = [] then Hashtbl.remove t dir
-    else Hashtbl.replace t dir fns
+    let fns = List.filter (fun (x, _) -> x <> id) fns in
+    if fns = [] then Hashtbl.remove t dir else Hashtbl.replace t dir fns
 
   let length t = Hashtbl.fold (fun _ v acc -> acc + List.length v) t 0
-
 end
 
 module Watchdog = struct
-
-  type t = {
-    t: (string, unit -> unit Lwt.t) Hashtbl.t;
-    d: Dispatch.t;
-  }
+  type t = { t : (string, unit -> unit Lwt.t) Hashtbl.t; d : Dispatch.t }
 
   let length t = Hashtbl.length t.t
+
   let dispatch t = t.d
 
   type hook = (string -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t
 
-  let empty (): t = {
-    t = Hashtbl.create 10;
-    d = Dispatch.empty ();
-  }
+  let empty () : t = { t = Hashtbl.create 10; d = Dispatch.empty () }
 
   let clear { t; d } =
     Hashtbl.fold (fun _dir stop acc -> acc >>= stop) t Lwt.return_unit
@@ -100,62 +101,64 @@ module Watchdog = struct
     Hashtbl.clear t;
     Dispatch.clear d
 
-  let watchdog t dir =
-    try Some (Hashtbl.find t dir) with Not_found -> None
+  let watchdog t dir = try Some (Hashtbl.find t dir) with Not_found -> None
 
   let start { t; d } ~dir listen =
     match watchdog t dir with
-    | Some _ -> assert (Dispatch.stats d ~dir <> 0); Lwt.return_unit
-    | None   ->
+    | Some _ ->
+        assert (Dispatch.stats d ~dir <> 0);
+        Lwt.return_unit
+    | None -> (
         (* Note: multiple threads can wait here *)
-        listen (fun file -> Dispatch.apply d ~dir ~file) >>= fun u ->
+        listen (fun file -> Dispatch.apply d ~dir ~file)
+        >>= fun u ->
         match watchdog t dir with
         | Some _ ->
             (* Note: someone else won the race, cancel our own thread
                to avoid avoid having too many wathdogs for [dir]. *)
             u ()
-        | None   ->
+        | None ->
             Log.debug (fun f -> f "Start watchdog for %s" dir);
             Hashtbl.add t dir u;
-            Lwt.return_unit
+            Lwt.return_unit)
 
   let stop { t; d } ~dir =
     match watchdog t dir with
-    | None      ->
+    | None ->
         assert (Dispatch.stats d ~dir = 0);
         Lwt.return_unit
     | Some stop ->
         if Dispatch.stats d ~dir <> 0 then (
           Log.debug (fun f -> f "Active allback are registered for %s" dir);
-          Lwt.return_unit
-        ) else (
+          Lwt.return_unit)
+        else (
           Log.debug (fun f -> f "Stop watchdog for %s" dir);
           Hashtbl.remove t dir;
-          stop ()
-        )
+          stop ())
 end
 
-type hook = int -> string -> (string -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t
+type hook =
+  int -> string -> (string -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t
 
 type t = {
-  mutable listen: int -> string -> (string -> unit Lwt.t) -> unit Lwt.t;
-  mutable stop  : unit -> unit Lwt.t;
-  watchdog      : Watchdog.t;
+  mutable listen : int -> string -> (string -> unit Lwt.t) -> unit Lwt.t;
+  mutable stop : unit -> unit Lwt.t;
+  watchdog : Watchdog.t;
 }
 
 let watchdog t = t.watchdog
 
-let hook t id dir f =
-  t.listen id dir f >|= fun () ->
-  t.stop
+let hook t id dir f = t.listen id dir f >|= fun () -> t.stop
 
 let create listen =
   let watchdog = Watchdog.empty () in
-  let t = {
-    listen = (fun _ _ _ -> Lwt.return_unit);
-    stop   = (fun _ -> Lwt.return_unit);
-    watchdog;
-  } in
+  let t =
+    {
+      listen = (fun _ _ _ -> Lwt.return_unit);
+      stop = (fun _ -> Lwt.return_unit);
+      watchdog;
+    }
+  in
   let listen id dir fn =
     let dir = realpath dir in
     let d = Watchdog.dispatch watchdog in