+++ packages/irmin-watcher.0.5.0/src/hook.ml
@@ -1,16 +1,17 @@
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire. All rights reserved.
    Distributed under the ISC license, see terms at the end of the file.
-   irmin-watcher 0.4.1
+   irmin-watcher 0.5.0
   ---------------------------------------------------------------------------*)
 
 open Lwt.Infix
 open Astring
 module Digests = Core.Digests
 
-let (/) = Filename.concat
+let ( / ) = Filename.concat
 
 let src = Logs.Src.create "irw-hook" ~doc:"Irmin watcher shared code"
+
 module Log = (val Logs.src_log src : Logs.LOG)
 
 let list_files kind dir =
@@ -21,33 +22,29 @@ let list_files kind dir =
     let d = List.filter kind d in
     let d = List.sort String.compare d in
     Lwt.return d
-  else
-  Lwt.return_nil
+  else Lwt.return_nil
 
 let directories dir =
-  list_files (fun f ->
-      try Sys.is_directory f with Sys_error _ -> false
-    ) dir
+  list_files (fun f -> try Sys.is_directory f with Sys_error _ -> false) dir
 
 let files dir =
-  list_files (fun f ->
-      try not (Sys.is_directory f) with Sys_error _ -> false
-    ) dir
+  list_files
+    (fun f -> try not (Sys.is_directory f) with Sys_error _ -> false)
+    dir
 
 let rec_files dir =
   let rec aux accu dir =
     directories dir >>= fun ds ->
-    files dir       >>= fun fs ->
-    Lwt_list.fold_left_s aux (fs @ accu) ds
+    files dir >>= fun fs -> Lwt_list.fold_left_s aux (fs @ accu) ds
   in
   aux [] dir
 
 let read_file ~prefix f =
   try
-    if not (Sys.file_exists f) || Sys.is_directory f then None
+    if (not (Sys.file_exists f)) || Sys.is_directory f then None
     else
-    let r = String.with_range ~first:(String.length prefix) f in
-    Some (r, Digest.file f)
+      let r = String.with_range ~first:(String.length prefix) f in
+      Some (r, Digest.file f)
   with ex ->
     Log.info (fun fm -> fm "read_file(%s): %a" f Fmt.exn ex);
     None
@@ -55,27 +52,26 @@ let read_file ~prefix f =
 let read_files dir =
   rec_files dir >|= fun new_files ->
   let prefix = dir / "" in
-  List.fold_left (fun acc f ->
-      match read_file ~prefix f with
-      | None   -> acc
-      | Some d -> Digests.add d acc
-    ) Digests.empty new_files
+  List.fold_left
+    (fun acc f ->
+      match read_file ~prefix f with None -> acc | Some d -> Digests.add d acc)
+    Digests.empty new_files
 
 type event = [ `Unknown | `File of string ]
 
-let rec poll n ~callback ~wait_for_changes dir files (event:event) =
-  ( match event with
+let rec poll n ~callback ~wait_for_changes dir files (event : event) =
+  (match event with
   | `Unknown -> read_files dir
-  | `File f  ->
+  | `File f -> (
       let prefix = dir / "" in
       let short_f = String.with_range ~first:(String.length prefix) f in
       let files = Digests.filter (fun (x, _) -> x <> short_f) files in
       match read_file ~prefix f with
-      | None   -> Lwt.return files
-      | Some d -> Lwt.return (Digests.add d files)
-  ) >>= fun new_files ->
-  Log.debug
-    (fun l -> l "files=%a new_files=%a" Digests.pp files Digests.pp new_files);
+      | None -> Lwt.return files
+      | Some d -> Lwt.return (Digests.add d files)))
+  >>= fun new_files ->
+  Log.debug (fun l ->
+      l "files=%a new_files=%a" Digests.pp files Digests.pp new_files);
   let diff = Digests.sdiff files new_files in
   let process () =
     if Digests.is_empty diff then Lwt.return_unit
@@ -91,10 +87,11 @@ let rec poll n ~callback ~wait_for_changes dir files (event:event) =
 let id = ref 0
 
 let v ~wait_for_changes ~dir callback =
-  let n = !id in incr id;
+  let n = !id in
+  incr id;
   read_files dir >|= fun files ->
-  Core.stoppable
-    (fun () -> poll n ~callback ~wait_for_changes dir files `Unknown)
+  Core.stoppable (fun () ->
+      poll n ~callback ~wait_for_changes dir files `Unknown)
 
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire