+++ packages/irmin-watcher.0.5.0/src/polling.mli
@@ -1,22 +1,24 @@
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire. All rights reserved.
    Distributed under the ISC license, see terms at the end of the file.
-   irmin-watcher 0.4.1
+   irmin-watcher 0.5.0
   ---------------------------------------------------------------------------*)
 
 (** Active polling backend for Irmin watchers.
 
-    {e 0.4.1 — {{:https://github.com/mirage/irmin-watcher }homepage}} *)
+    {e 0.5.0 — {{:https://github.com/mirage/irmin-watcher} homepage}} *)
 
 open Core
 
-type event = [ `Unknown | `File of string ]
-(** The type for change event. *)
+val with_delay : float -> t
+(** [with_delay delay id p f] is the hook calling [f] everytime a sub-path of
+    [p] is modified. Return a function to call to remove the hook. Active
+    polling is done every [delay] seconds. *)
 
-val v: wait_for_changes:(unit -> event Lwt.t) -> dir:string -> Watchdog.hook
-(** [v ~wait_for_changes ~dir] is the watchdog hook using
-    [wait_for_changes] to detect filesystem updates in the directory
-    [dir]. The polling implemention just calls [Lwt_unix.sleep]. *)
+val v : t Lazy.t
+(** [v] is [with_delay !default_polling_time]. *)
+
+val mode : [ `Polling ]
 
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Thomas Gazagnaire